Avoid errant assert for visitNode empty array case
The `visitNode` function can optionally return an array of replacement nodes. If the array is empty (indicating no nodes should be used), the `extractSingleNode` function will correctly return undefined via `singleOrUndefined`. However, the location of the assert call within `visitNode` assumes that the output node of the visit (`visitedNode` variable) is defined. This results in an assertion failure when TypeScript is in "development" mode even though the returned undefined value is valid.